requirements.txt
python-dotenv
slack_bolt
langchain
openai
pytest
pre-commit
black
flake8
pandas
matplotlib
plotly
scipy
scikit-learn
pdfminer
pdfminer.six
rich
flask
notion_client
feat_copybot_PDF
PyGithub